New CLI for hydrology-render.py issue-78
