+++ hydrology2.py
@@ -2,7 +2,7 @@
 
 import argparse
 
-from TerrainHydrology.ModelIO import Export
+from TerrainHydrology.ModelIO import Export, Render
 
 def export(args: argparse.Namespace) -> None:
     if args.nodeOutput is not None:
@@ -15,6 +15,9 @@ def export(args: argparse.Namespace) -> None:
         Export.writeDownstreamEdgeShapefile(args.inputFile, args.latitude, args.longitude, args.downstreamEdgeOutput)
     if args.riverOutput is not None:
         Export.writeRiverShapefile(args.inputFile, args.latitude, args.longitude, args.riverOutput)
+
+def render(args: argparse.Namespace) -> None:
+    Render.renderDEM(args.inputFile, args.latitude, args.longitude, args.outputResolution, args.num_procs, args.outputDir, args.extremeMemory)
 
 parser = argparse.ArgumentParser(
     description='Terrain system based on Genevaux et al., "Terrain Generation Using Procedural Models Based on Hydrology", ACM Transactions on Graphics, 2013'
@@ -98,6 +101,66 @@ parser_export.add_argument(
     required=False
 )
 parser_export.set_defaults(func=export)
+
+parser_render = subparsers.add_parser('render', help='render help')
+parser_render.add_argument(
+    '-i',
+    '--input',
+    help='The file that contains the data model you wish to render',
+    dest='inputFile',
+    metavar='output/data',
+    required=True
+)
+parser_render.add_argument(
+    '--lat',
+    metavar='-43.2',
+    help='Center latitude for the output GeoTIFF',
+    type=float,
+    dest='latitude',
+    required=True
+)
+parser_render.add_argument(
+    '--lon',
+    metavar='-103.8',
+    help='Center longitude for the output GeoTIFF',
+    type=float,
+    dest='longitude',
+    required=True
+)
+parser_render.add_argument(
+    '-ro',
+    '--output-resolution',
+    metavar='1000',
+    help='The number of pixels/samples on each side of the output raster',
+    dest='outputResolution',
+    type=int,
+    required=True
+)
+parser_render.add_argument(
+    '--num-procs',
+    metavar='4',
+    help='The number of processes/threads to use in rendering. This should be the number of cores you have on your system.',
+    dest='num_procs',
+    type=int,
+    default=4,
+    required=False
+)
+parser_render.add_argument(
+    '-o',
+    '--output-dir',
+    help='Folder in which to dump all the data generated by this program (including debug data)',
+    dest='outputDir',
+    metavar='output/',
+    required=True
+)
+parser_render.add_argument(
+    '--extreme-memory',
+    help='An experimental method of conserving memory when rendering large terrains',
+    dest='extremeMemory',
+    action='store_true',
+    required=False
+)
+parser_render.set_defaults(func=render)
 
 args = parser.parse_args()
 args.func(args)
